extract wave read support routes
